deletion from dynamic route complete

+++ todo-list-v2/app.js
@@ -3,6 +3,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require('mongoose');
+const _ = require("lodash");
 
 const app = express();
 
@@ -114,15 +115,31 @@ else
 app.post("/delete",function(req,res)
 {
  const checkedItemId = req.body.checkbox;
-
- Item.findByIdAndRemove(checkedItemId,function(err)
-{
-  if(!err)
+ const listName = req.body.listName;
+ if(listName == "Today")
+ {
+   Item.findByIdAndRemove(checkedItemId,function(err)
   {
-    console.log("Successfully deleted the item!");
-  }
-});
-res.redirect("/");
+    if(!err)
+    {
+      console.log("Successfully deleted the item!");
+    }
+  });
+  res.redirect("/");
+ }
+ else
+ {
+   List.findOneAndUpdate({name:listName},{$pull: {items:{_id:checkedItemId}}},function(err, foundList)
+ {
+   if(!err)
+   {
+     res.redirect("/"+listName);
+   }
+ });
+
+ }
+
+
 
 });
 
@@ -147,7 +164,7 @@ app.post("/:customListName",function(req,res)
 app.get("/:customListName", function(req,res)
 {
 
-  const customListName = req.params.customListName
+  const customListName = _.capitalize(req.params.customListName)
 
    List.findOne({name:customListName},function(err,foundList)
    {
